requirements.txt
discord.py
psycopg2
python-dotenv